+++ tt/tests/unit/expressions/_helpers.py
@@ -10,6 +10,14 @@ class ExpressionTestCase(unittest.TestCase):
 
     """An extended TestCase with helpers for testing expressions."""
 
+    def assert_is_cnf(self, expr_str):
+        """Assert an expression (given as a string) is in cnf form."""
+        self.assertTrue(BooleanExpression(expr_str).is_cnf)
+
+    def assert_not_cnf(self, expr_str):
+        """Assert an expression (given as a string) is not in cnf form."""
+        self.assertFalse(BooleanExpression(expr_str).is_cnf)
+
     def helper_test_evaluate(self, expr, expected_result=None, **kwargs):
         """Helper for testing the evaluation of expressions.